+++ locals.tf
@@ -8,5 +8,5 @@ locals {
     commandToExecute = "/bin/bash -c \"[[ $(awk -F'=' '/^ID/ { print $2 }' /etc/os-release 2>/dev/null) =~ 'debian' ]] && apt update && apt install -y gpg || true\""
   }
 
-  diag_name = format("%s-%s", element(split("/", var.vm_id), 8), var.vm_extension_name_suffix)
+  diag_name = lower(format("%s-%s", element(split("/", var.vm_id), 8), var.vm_extension_name_suffix))
 }