+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Breaking
 Updated
   * AZ-273: Module now compatible terraform `v0.13+`
 
+Changed
+  * AZ-398: Force lowercase on default generated name
+
 # v3.0.0 - 2020-07-22
 
 Breaking