+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Updated
   * AZ-532: Revamp README with latest `terraform-docs` tool
+  * AZ-530: Cleanup module, fix linter errors
 
 # v4.0.0 - 2021-01-21